Split the scene_builder module into smaller files.
