--- agents_playground/project/rules/agents-playground.code-workspace
@@ -1,12 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../../.."
-		}
-	],
-	"settings": {
-		"cSpell.words": [
-			"funcs"
-		]
-	}
-}
